Refactor Window Covering cluster: move ZAP APIs to CodegenIntegration and standardize setters   - Move ZAP-dependent conversion functions (LiftToPercent100ths, Percent100thsToLift, TiltToPercent100ths, Percent100thsToTilt) from WindowCoveringCluster.cpp to CodegenIntegration.cpp   - Remove #include <Accessors.h> from core cluster code   - Standardize setters (SetConfigStatus, SetOperationalStatus, SetEndProductType, SetSafetyStatus, SetType, SetTargetPosition*) to use SetAttributeValue helper   - Move WC_PERCENT100THS_* constants into WindowCovering namespace   - Remove unused macros (CHECK_BOUNDS_*, FAKE_MOTION_DELAY_MS) and #include <string.h>   - Fix variable shadowing namespace in FindClusterOnEndpoint   - Fix CodegenIntegration.h to include WindowCoveringCluster.h directly
